+++ examples/simple-vpc/main.tf
@@ -31,11 +31,6 @@ provider "aws" {
   }
 }
 
-data "aws_security_group" "default" {
-  name   = "default"
-  vpc_id = module.vpc.vpc_id
-}
-
 module "vpc" {
   source = "../../"