+++ examples/vpc-flow-logs/versions.tf
@@ -2,6 +2,7 @@ terraform {
   required_version = ">= 0.12.7, < 0.14"
 
   required_providers {
-    aws = ">= 2.68, < 4.0"
+    aws    = ">= 2.68, < 4.0"
+    random = ">= 2"
   }
 }